DABOM-129 rename: 로그인 관련 클래스를 global 패키지로 이전
