+++ src/main/java/com/project/customer/application/SignInService.java
@@ -1,5 +1,6 @@
 package com.project.customer.application;
 
+import com.project.global.auth.JwtTokenUtil;
 import org.springframework.stereotype.Service;
 
 import com.project.customer.core.Role;
