+++ src/main/java/com/project/global/auth/LoginInterceptor.java
@@ -1,4 +1,4 @@
-package com.project.customer.application;
+package com.project.global.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
